feat: add Select component into schema components
