+++ packages/client/src/schema-component/antd/select/index.md
@@ -5,4 +5,17 @@ group:
   path: /schema-components
 ---
 
-# Select <Badge>待定</Badge>
+# Select 
+
+## Nodes
+
+* Select
+* Select. Object
+
+## Examples
+
+### Select
+
+<code src="./demos/demo1.tsx" />
+
+<code src="./demos/demo2.tsx" />
